src/requirements.txt
Flask
google-api-python-client
flask-sqlalchemy
beautifulsoup4
urllib2